abstraction of servlets with service level assignment: demo-implementation for 'hello' servlet
