+++ src/org/loklak/server/AbstractAPIHandler.java
@@ -20,6 +20,7 @@
 package org.loklak.server;
 
 import java.io.IOException;
+import java.io.PrintWriter;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -27,6 +28,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.json.JSONObject;
+import org.loklak.http.RemoteAccess;
 
 @SuppressWarnings("serial")
 public abstract class AbstractAPIHandler extends HttpServlet implements APIHandler {
@@ -40,32 +42,52 @@ public abstract class AbstractAPIHandler extends HttpServlet implements APIHandl
     public AbstractAPIHandler(String[] serverProtocolHostStub) {
         this.serverProtocolHostStub = serverProtocolHostStub;
     }
-    
-    @Override
-    public JSONObject Service(JSONObject call) {
-        return null;
-    }
+
     @Override
     public String[] getServerProtocolHostStub() {
         return this.serverProtocolHostStub;
     }
 
     @Override
-    public APIServiceLevel getDefaultServiceLevel() {
-        return APIServiceLevel.PUBLIC;
-    }
+    public abstract APIServiceLevel getDefaultServiceLevel();
 
     @Override
-    public APIServiceLevel getCustomServiceLevel(HTTPAccount customer) {
-        return null;
-    }
+    public abstract APIServiceLevel getCustomServiceLevel(Authorization auth);
+    
+    @Override
+    public abstract JSONObject service(JSONObject call);
 
     @Override
-    public void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         doGet(request, response);
     }
     
     @Override
-    abstract public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException;
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        RemoteAccess.Post post = RemoteAccess.evaluate(request);
+        
+        // manage DoS
+        if (post.isDoS_blackout()) {response.sendError(503, "your request frequency is too high"); return;}
+        
+        String callback = post.get("callback", "");
+        boolean jsonp = callback != null && callback.length() > 0;
+        
+        post.setResponse(response, "application/javascript");
+        
+        // transform keys to JSONObject
+        JSONObject args = new JSONObject();
+        for (String key: post.getKeys()) args.put(key, post.get(key, "")); // all object values are string
+        
+        // generate json
+        JSONObject json = service(args);
 
+        // write json
+        response.setCharacterEncoding("UTF-8");
+        PrintWriter sos = response.getWriter();
+        if (jsonp) sos.print(callback + "(");
+        sos.print(json.toString(2));
+        if (jsonp) sos.println(");");
+        sos.println();
+        post.finalize();
+    }
 }
